requirements.txt
fastapi
fastapi-cors
uvicorn[standard]
google-generativeai